toggleswitch-LED-SDCard
Code that uses a toggleswitch, LED, and MicroSD card (writes data to the SD card). See here for more examples on how to read/write data to/from the MicroSD card https://create.arduino.cc/projecthub/electropeak/sd-card-module-with-arduino-how-to-read-write-data-37f390 You could modify to use the toggleswitch and using if statements to log data based on toggling the switch.